+++ src/tasks/start/utils/backend/ganache.ts
@@ -3,11 +3,7 @@ import { promisify } from 'util'
 import { BuidlerRuntimeEnvironment } from '@nomiclabs/buidler/types'
 import { BuidlerPluginError } from '@nomiclabs/buidler/plugins'
 import { logBack } from '../logger'
-import {
-  aragenGasLimit,
-  aragenMnemonic,
-  testnetPort
-} from '~/src/aragon-params'
+import { aragenGasLimit, aragenMnemonic, testnetPort } from '~/src/params'
 
 // There's an issue with how web3 exports its typings that conflicts with
 // ganache-core imports of those typings. Follow https://github.com/trufflesuite/ganache-core/issues/465
